refactor: 테이블 컴포넌트 파일 정리
