+++ shared/@common/ui/Button/ButtonComponent.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Table from '@/shared/@common/ui/Table/TableBody';
+import Table from '@/shared/@common/ui/Table/3';
 import Button from '@/shared/@common/ui/Button/Button';
 
 const ButtonComponent = () => {
